+++ src/preset/bun/providers/config-source.ts
@@ -1,0 +1,6 @@
+/* eslint-disable require-jsdoc, jsdoc/require-jsdoc */
+import { ConfigSource, createConfigSource } from '../../../config';
+
+export function provideConfigSource(): ConfigSource {
+  return createConfigSource(Bun.env);
+}
