+++ src/preset/bun/providers/fetch-middleware.ts
@@ -1,0 +1,7 @@
+/* eslint-disable require-jsdoc, jsdoc/require-jsdoc */
+import { Middleware } from '../../../http';
+
+export function provideFetchMiddleware(): Middleware[] {
+  // @todo добавить логирование ошибок?
+  return [];
+}
