+++ src/preset/bun/providers/service-routes.ts
@@ -1,0 +1,12 @@
+/* eslint-disable require-jsdoc, jsdoc/require-jsdoc */
+import { ServerHandler } from '../../server/types';
+import { getStatsHandler } from '../utils/get-stats-handler';
+import { getHealthCheck } from '../../server/utils/get-health-check';
+
+export function provideServiceRoutes(): Array<[string, ServerHandler]> {
+  return [
+    // служебные маршруты (без промежуточных слоев)
+    ['/healthcheck', getHealthCheck()],
+    ['/stats', getStatsHandler()],
+  ];
+}
